+++ NekrasovskyAPP/NekrasovskyAPP/Models/FillingWarehouse.cs
@@ -22,7 +22,7 @@ namespace NekrasovskyAPP.Models
 
         [Required]
         [Column(TypeName = "integer")]
-        public int Quantity { get; set; } = 0;
+        public double Quantity { get; set; } = 0;
 
         [Column(TypeName = "varchar(20)")]
         public string? MeasuringType { get; set; }
