+++ server/server/Models/ResponsibilityStock.cs
@@ -8,7 +8,7 @@ namespace server.Models
         public int ItemId { get; set; }
         public string Name { get; set; } = string.Empty;
         public string MeasuringUnit { get; set; } = string.Empty;
-        public int TotalQuantity { get; set; }
+        public double TotalQuantity { get; set; }
         public List<ResponsibilityWarehouseStock> Warehouses { get; set; } = new();
     }
 
@@ -16,7 +16,7 @@ namespace server.Models
     {
         public int WarehouseId { get; set; }
         public string WarehouseName { get; set; } = string.Empty;
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
         public string MeasuringType { get; set; } = string.Empty;
     }
 }
